refactor: decoupling typescript from core
