+++ packages/language-service/src/languageFeatures/validation.ts
@@ -118,7 +118,7 @@ type CacheMap = Map<
 		string,
 		{
 			documentVersion: number,
-			tsProjectVersion: string | undefined,
+			projectVersion: string | undefined,
 			errors: vscode.Diagnostic[] | undefined | null,
 		}
 	>
@@ -168,7 +168,7 @@ export function register(context: ServiceContext) {
 			syntax_rules: { errors: [] },
 			format_rules: { errors: [] },
 		}).get(uri)!;
-		const newSnapshot = context.host.getScriptSnapshot(context.env.uriToFileName(uri));
+		const newSnapshot = context.core.host.getScriptSnapshot(context.env.uriToFileName(uri));
 
 		let updateCacheRangeFailed = false;
 		let errorsUpdated = false;
@@ -262,10 +262,10 @@ export function register(context: ServiceContext) {
 
 					const pluginCache = cacheMap.get(ruleId) ?? cacheMap.set(ruleId, new Map()).get(ruleId)!;
 					const cache = pluginCache.get(lintDocument.uri);
-					const tsProjectVersion = (ruleType === RuleType.Semantic) ? context.core.typescript.languageServiceHost.getProjectVersion?.() : undefined;
+					const projectVersion = (ruleType === RuleType.Semantic) ? context.core.host.getProjectVersion?.() : undefined;
 
 					if (ruleType === RuleType.Semantic) {
-						if (cache && cache.documentVersion === lintDocument.version && cache.tsProjectVersion === tsProjectVersion) {
+						if (cache && cache.documentVersion === lintDocument.version && cache.projectVersion === projectVersion) {
 							return cache.errors;
 						}
 					}
@@ -321,7 +321,7 @@ export function register(context: ServiceContext) {
 					pluginCache.set(lintDocument.uri, {
 						documentVersion: lintDocument.version,
 						errors,
-						tsProjectVersion,
+						projectVersion,
 					});
 
 					return errors;
@@ -364,10 +364,10 @@ export function register(context: ServiceContext) {
 					const serviceId = Object.keys(context.services).find(key => context.services[key] === service)!;
 					const serviceCache = cacheMap.get(serviceId) ?? cacheMap.set(serviceId, new Map()).get(serviceId)!;
 					const cache = serviceCache.get(document.uri);
-					const tsProjectVersion = api === 'provideSemanticDiagnostics' ? context.core.typescript.languageServiceHost.getProjectVersion?.() : undefined;
+					const projectVersion = api === 'provideSemanticDiagnostics' ? context.core.host.getProjectVersion?.() : undefined;
 
 					if (api === 'provideSemanticDiagnostics') {
-						if (cache && cache.documentVersion === document.version && cache.tsProjectVersion === tsProjectVersion) {
+						if (cache && cache.documentVersion === document.version && cache.projectVersion === projectVersion) {
 							return cache.errors;
 						}
 					}
@@ -399,7 +399,7 @@ export function register(context: ServiceContext) {
 					serviceCache.set(document.uri, {
 						documentVersion: document.version,
 						errors,
-						tsProjectVersion,
+						projectVersion,
 					});
 
 					return errors;
